LIME-5: * changed clsx to getValidClassName helper
